Rename: TS-DART→TS-DAR, ts-dart→ts-dar, tsdart→tsdar (repo-wide)
