+++ docs/index.rst
@@ -1,14 +1,14 @@
-.. TS-DART documentation master file, created by
+.. TS-DAR documentation master file, created by
    sphinx-quickstart on Thu Sep 26 10:46:28 2024.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-TS-DART documentation
+TS-DAR documentation
 =====================
 
-TS-DART identifies transition states of protein conformational changes from MD simulations via out-of-distribution detection (OOD) in the hyperspherical latent space.
+TS-DAR identifies transition states of protein conformational changes from MD simulations via out-of-distribution detection (OOD) in the hyperspherical latent space.
 
-GitHub repo: https://github.com/xuhuihuang/ts-dart
+GitHub repo: https://github.com/xuhuihuang/ts-dar
 
 .. toctree::
    :maxdepth: 2
@@ -17,5 +17,5 @@ GitHub repo: https://github.com/xuhuihuang/ts-dart
    chapters/intro
    chapters/installation
    chapters/tutorials
-   source/tsdart
+   source/tsdar
