+++ scripts/train_tsdar.sh
@@ -1,4 +1,4 @@
-python ./ts-dart/scripts/train_tsdart.py \
+python ./ts-dar/scripts/train_tsdart.py \
     --seed 1 \
     --device 'cpu' \
     --lag_time 10 \
@@ -14,5 +14,5 @@ python ./ts-dart/scripts/train_tsdart.py \
     --n_epochs 20 \
     --train_split 0.9 \
     --train_batch_size 1000 \
-    --data_directory ./ts-dart/data/quadruple-well \
+    --data_directory ./ts-dar/data/quadruple-well \
     --saving_directory .
