+++ tsdar/utils.py
